+++ R/utils.R
@@ -92,6 +92,29 @@ downloadIfNeeded <- function(file, link, gz=TRUE) {
     }
 }
 
+#' Collapse duplicated rows based on a column of a data frame
+#'
+#' @param df Data frame
+#' @param column Character: name of the column with elements to indicate which
+#' rows to collapse
+#'
+#' @importFrom plyr ldply
+#'
+#' @return Data frame with duplicated rows collapsed
+#' @keywords internal
+collapseDuplicatedRows <- function(df, column) {
+    cluster <- split(seq(df[[column]]), df[[column]])
+    collapseRowInfo <- function(thisCluster, df) {
+        tmp <- df[thisCluster, ]
+        apply(tmp, 2, function(x) paste(unique(x), collapse=", "))
+    }
+    collapsed <- pblapply(cluster, collapseRowInfo, df)
+    collapsed <- ldply(collapsed, .id=column)
+    return(collapsed)
+}
+
+# l1000perturbations object ----------------------------------------------------
+
 #' Subset an \code{l1000perturbations} object
 #'
 #' @param x \code{l1000perturbations} object
@@ -101,7 +124,7 @@ downloadIfNeeded <- function(file, link, gz=TRUE) {
 #' @export
 `[.l1000perturbations` <- function(x, ...) {
     out <- unclass(x)
-    out <- `[`(out, ...)
+    out <- `[`(out, ..., drop=FALSE)
 
     # Inherit the same attributes
     attrs <- attributes(x)